fix: Quote validators
